# openapi.json
{"openapi":"3.1.0","info":{"title":"Rationale Generator API","version":"0.1.0"},"paths":{"/gemini/analyze-with-rationale":{"post":{"tags":["Gemini"],"summary":"Analyze With Rationale","operationId":"analyze_with_rationale_gemini_analyze_with_rationale_post","parameters":[{"name":"X-GEMINI-API-KEY","in":"header","required":true,"schema":{"type":"string","title":"X-Gemini-Api-Key"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_analyze_with_rationale_gemini_analyze_with_rationale_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/gemini/analyze-image-only":{"post":{"tags":["Gemini"],"summary":"Analyze Image","operationId":"analyze_image_gemini_analyze_image_only_post","parameters":[{"name":"X-GEMINI-API-KEY","in":"header","required":true,"schema":{"type":"string","title":"X-Gemini-Api-Key"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_analyze_image_gemini_analyze_image_only_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/login":{"post":{"tags":["Admin"],"summary":"Admin Login","operationId":"admin_login_admin_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/clients/create":{"post":{"tags":["Clients"],"summary":"Create Client","operationId":"create_client_admin_clients_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/clients/":{"get":{"tags":["Clients"],"summary":"Get All Clients","operationId":"get_all_clients_admin_clients__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/clients/{client_id}":{"get":{"tags":["Clients"],"summary":"Get Client By Id","operationId":"get_client_by_id_admin_clients__client_id__get","parameters":[{"name":"client_id","in":"path","required":true,"schema":{"type":"integer","title":"Client Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/usage/create":{"post":{"tags":["Usage"],"summary":"Create Usage","operationId":"create_usage_usage_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsageCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sheets":{"post":{"tags":["sheets"],"summary":"Create Sheet","description":"Create a new sheet for the current user","operationId":"create_sheet_api_sheets_post","parameters":[{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"X-User-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SheetCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SheetResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["sheets"],"summary":"Get All Sheets","description":"Get all sheets for the current user, optionally filtered by date","operationId":"get_all_sheets_api_sheets_get","parameters":[{"name":"date_filter","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by upload date (YYYY-MM-DD)","title":"Date Filter"},"description":"Filter by upload date (YYYY-MM-DD)"},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"X-User-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SheetListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sheets/{sheet_id}":{"get":{"tags":["sheets"],"summary":"Get Sheet","description":"Get a specific sheet by ID","operationId":"get_sheet_api_sheets__sheet_id__get","parameters":[{"name":"sheet_id","in":"path","required":true,"schema":{"type":"integer","title":"Sheet Id"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"X-User-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SheetResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["sheets"],"summary":"Delete Sheet","description":"Delete a sheet and all its row rationales","operationId":"delete_sheet_api_sheets__sheet_id__delete","parameters":[{"name":"sheet_id","in":"path","required":true,"schema":{"type":"integer","title":"Sheet Id"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"X-User-Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sheets/{sheet_id}/processed-rows":{"put":{"tags":["sheets"],"summary":"Update Processed Rows","description":"Update the processed rows for a sheet","operationId":"update_processed_rows_api_sheets__sheet_id__processed_rows_put","parameters":[{"name":"sheet_id","in":"path","required":true,"schema":{"type":"integer","title":"Sheet Id"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"X-User-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"integer"},"title":"Processed Rows"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SheetResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sheets/rationales":{"post":{"tags":["sheets"],"summary":"Create Row Rationale","description":"Create or update a row rationale","operationId":"create_row_rationale_api_sheets_rationales_post","parameters":[{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"X-User-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RowRationaleCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RowRationaleResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sheets/rationales/{sheet_id}/{row_index}":{"get":{"tags":["sheets"],"summary":"Get Row Rationale","description":"Get rationale for a specific row","operationId":"get_row_rationale_api_sheets_rationales__sheet_id___row_index__get","parameters":[{"name":"sheet_id","in":"path","required":true,"schema":{"type":"integer","title":"Sheet Id"}},{"name":"row_index","in":"path","required":true,"schema":{"type":"integer","title":"Row Index"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"X-User-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RowRationaleResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sheets/rationales/sheet/{sheet_id}":{"get":{"tags":["sheets"],"summary":"Get All Rationales For Sheet","description":"Get all rationales for a specific sheet","operationId":"get_all_rationales_for_sheet_api_sheets_rationales_sheet__sheet_id__get","parameters":[{"name":"sheet_id","in":"path","required":true,"schema":{"type":"integer","title":"Sheet Id"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"X-User-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RowRationaleResponse"},"title":"Response Get All Rationales For Sheet Api Sheets Rationales Sheet  Sheet Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sheets/rationales/{rationale_id}":{"put":{"tags":["sheets"],"summary":"Update Row Rationale","description":"Update a row rationale","operationId":"update_row_rationale_api_sheets_rationales__rationale_id__put","parameters":[{"name":"rationale_id","in":"path","required":true,"schema":{"type":"integer","title":"Rationale Id"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"X-User-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RowRationaleUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RowRationaleResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["sheets"],"summary":"Delete Row Rationale","description":"Delete a row rationale","operationId":"delete_row_rationale_api_sheets_rationales__rationale_id__delete","parameters":[{"name":"rationale_id","in":"path","required":true,"schema":{"type":"integer","title":"Rationale Id"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"X-User-Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/login":{"post":{"tags":["Client Auth"],"summary":"Client Login","operationId":"client_login_api_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Health","operationId":"health__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AdminLogin":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"AdminLogin"},"Body_analyze_image_gemini_analyze_image_only_post":{"properties":{"image":{"type":"string","format":"binary","title":"Image"}},"type":"object","required":["image"],"title":"Body_analyze_image_gemini_analyze_image_only_post"},"Body_analyze_with_rationale_gemini_analyze_with_rationale_post":{"properties":{"trade_data":{"type":"string","title":"Trade Data"},"image":{"type":"string","format":"binary","title":"Image"},"plan_type":{"anyOf":[{"$ref":"#/components/schemas/PlanType"},{"type":"null"}]},"prompt":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Prompt"}},"type":"object","required":["trade_data","image"],"title":"Body_analyze_with_rationale_gemini_analyze_with_rationale_post"},"ClientCreate":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"ClientCreate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PlanType":{"type":"string","enum":["Equity","Commodity","Options","Derivatives"],"title":"PlanType"},"RowRationaleCreate":{"properties":{"sheet_id":{"type":"integer","title":"Sheet Id"},"row_index":{"type":"integer","title":"Row Index"},"rationale_text":{"type":"string","title":"Rationale Text"},"rationale_result":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Rationale Result"},"image_preview":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Preview"},"editable_rationale":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Editable Rationale"}},"type":"object","required":["sheet_id","row_index","rationale_text"],"title":"RowRationaleCreate"},"RowRationaleResponse":{"properties":{"id":{"type":"integer","title":"Id"},"sheet_id":{"type":"integer","title":"Sheet Id"},"row_index":{"type":"integer","title":"Row Index"},"rationale_text":{"type":"string","title":"Rationale Text"},"rationale_result":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Rationale Result"},"image_preview":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Preview"},"editable_rationale":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Editable Rationale"},"generated_date":{"type":"string","format":"date-time","title":"Generated Date"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","sheet_id","row_index","rationale_text","generated_date","created_at","updated_at"],"title":"RowRationaleResponse"},"RowRationaleUpdate":{"properties":{"rationale_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Rationale Text"},"rationale_result":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Rationale Result"},"image_preview":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Preview"},"editable_rationale":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Editable Rationale"}},"type":"object","title":"RowRationaleUpdate"},"SheetCreate":{"properties":{"file_name":{"type":"string","title":"File Name"},"upload_date":{"type":"string","title":"Upload Date"},"rows_data":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Rows Data"},"processed_rows":{"anyOf":[{"items":{"type":"integer"},"type":"array"},{"type":"null"}],"title":"Processed Rows","default":[]}},"type":"object","required":["file_name","upload_date","rows_data"],"title":"SheetCreate"},"SheetListResponse":{"properties":{"sheets":{"items":{"$ref":"#/components/schemas/SheetResponse"},"type":"array","title":"Sheets"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["sheets","total"],"title":"SheetListResponse"},"SheetResponse":{"properties":{"id":{"type":"integer","title":"Id"},"client_id":{"type":"integer","title":"Client Id"},"file_name":{"type":"string","title":"File Name"},"upload_date":{"type":"string","title":"Upload Date"},"rows_data":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Rows Data"},"processed_rows":{"items":{"type":"integer"},"type":"array","title":"Processed Rows"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","client_id","file_name","upload_date","rows_data","processed_rows","created_at","updated_at"],"title":"SheetResponse"},"UsageCreate":{"properties":{"client_id":{"type":"integer","title":"Client Id"},"usage":{"items":{"$ref":"#/components/schemas/UsageItem"},"type":"array","title":"Usage"}},"type":"object","required":["client_id","usage"],"title":"UsageCreate"},"UsageItem":{"properties":{"action":{"type":"string","title":"Action"},"tokens_used":{"type":"integer","title":"Tokens Used"}},"type":"object","required":["action","tokens_used"],"title":"UsageItem"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}